BookingService/src/requirements.txt
Flask
mysql-connector-python
python-dotenv
flasgger
flask-jwt-extended